+++ packages/adapters/src/serverEngines/hono.js
@@ -80,7 +80,7 @@ class HonoEngine {
 
         let port = options.port;
 
-        server.once('ready', async () => {
+        server.once('after:Final', async () => {
             return new Promise((resolve, reject) => {
                 server.httpServer.listen(port, function (err) {
                     if (err) {
@@ -96,8 +96,8 @@ class HonoEngine {
                         host = address.address;
                     }
 
-                    server.host = `${host}:${address.port}`;
-                    server.port = address.port;
+                    server.httpHost = `${host}:${address.port}`;
+                    server.httpPort = address.port;
 
                     server.log('info', `A koa http service is listening on port [${address.port}] ...`);
                     resolve();