+++ packages/tester/src/login/password.js
@@ -11,7 +11,6 @@ import { get } from '@kitmi/utils';
  * @property {String} [loginOptions.tokenKey] - The token key path, default "token".
  * @property {Object} [loginOptions.query] - The query parameters.
  * @property {Object} [loginOptions.headers] - Extra headers.
- * @private
  */
 const passwordAuth = async (client, loginOptions) => {
     const usernameField = loginOptions.usernameField || 'username';